[py] Rounded pause duration (#6300)
When specifying a fraction of a second, the multiplication of 1000 causes a float value to be sent to the endpoint which causes an error

+++ py/selenium/webdriver/common/actions/pointer_input.py
@@ -54,7 +54,7 @@ class PointerInput(InputDevice):
         self.add_action({"type": "pointerCancel"})
 
     def create_pause(self, pause_duration):
-        self.add_action({"type": "pause", "duration": pause_duration * 1000})
+        self.add_action({"type": "pause", "duration": int(pause_duration * 1000)})
 
     def encode(self):
         return {"type": self.type,
